requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
dj-database-url==2.1.0
requests==2.31.0